requirements.txt
annotated-types==0.7.0
anyio==4.3.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.5
crontab==1.0.1
cymem==2.0.8
distlib==0.3.9
ecdsa==0.19.0
eng_to_ipa==0.0.2
exceptiongroup==1.2.2
fastapi==0.110.3
filelock==3.13.4
freezegun==1.5.1
fsspec==2024.3.1
gensim==4.3.3
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.3.0
MarkupSafe==2.1.5
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.4.2
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
packaging==24.1
passlib==1.7.4
pbr==6.0.0
pillow==10.3.0
platformdirs==4.2.2
preshed==3.0.9
pyasn1==0.6.1
pydantic==2.7.4
pydantic_core==2.18.4
PyMySQL==1.1.1
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
redis==5.0.4
regex==2024.4.28
requests==2.31.0
rq==1.16.1
rq-scheduler==0.13.1
rsa==4.9
scikit-learn==1.5.2
scipy==1.13.1
setuptools==59.6.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
srsly==2.4.8
starlette==0.37.2
stevedore==5.2.0
sympy==1.13.3
textblob==0.18.0.post0
thinc==8.2.5
threadpoolctl==3.5.0
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.66.5
triton==2.2.0
typer==0.9.4
typing_extensions==4.11.0
urllib3==2.2.3
uvicorn==0.29.0
virtualenv==20.25.3
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.1
wasabi==1.1.3
weasel==0.3.4
